requirements.txt
altair==5.4.0
attrs==24.1.0
Automat==22.10.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
configparser==7.0.0
constantly==23.10.4
contourpy==1.2.1
coverage==7.6.1
coveralls==4.0.1
cryptography==43.0.0
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
docopt==0.6.2
emoji==2.12.1
entrypoints==0.4
executing==2.0.1
Faker==26.3.0
fastjsonschema==2.20.0
favicon==0.7.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
gensim==4.3.3
genson==1.3.0
gitdb==4.0.11
GitPython==3.1.43
htbuilder==0.6.2
huggingface-hub==0.24.5
hyperlink==21.0.0
hypothesis==6.108.8
idna==3.7
incremental==24.7.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mock==5.1.0
more-itertools==10.4.0
narwhals==1.3.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nlp_id==0.1.15.0
nltk==3.8.1
numpy==1.24.3
packaging==24.1
pandas==2.0.3
parso==0.8.4
pdfkit==1.0.0
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pymdown-extensions==10.9
pyOpenSSL==24.2.1
pyparsing==3.1.2
PySastrawi==1.2.0
pytest==7.3.1
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
scikit-learn==1.2.2
scipy==1.13.1
seaborn==0.13.2
setuptools==67.7.1
simplejson==3.19.2
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sortedcontainers==2.4.0
soupsieve==2.5
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
streamlit==1.37.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.6
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-shadcn-ui==0.1.18
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tabpy==2.10.0
tenacity==8.5.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.2
validators==0.33.0
watchdog==4.0.2
wcwidth==0.2.13
wget==3.2
wheel==0.40.0
wordcloud==1.9.3
wrapt==1.16.0
zope.interface==6.4.post2
build==0.8.0
